- Add visualisation of friends - Import Books and Filters from the database of a friend   - import Publishers, Authors, and Categories required - A friend can be deleted (Deleted a friend deleted their book, filters, loans)

+++ app/src/main/java/flq/projectbooks/data/libraries/LoanLibrary.java
@@ -103,7 +103,7 @@ public class LoanLibrary implements Serializable {
         return loans;
     }
 
-    public void deleteLoanById(int id) {
+    public void deleteLoanById(long id) {
         for (int j = 0; j < loanList.size(); j++) {
             if (loanList.get(j).getId() == id) {
                 //Remove from database
